接口优化

+++ app/src/main/java/com/example/lw/learnkotlin/bean/Coordinates.kt
@@ -4,4 +4,4 @@ package com.example.lw.learnkotlin.bean
  * Created on 2018/11/23.
  * @author Alan
  */
-data class Coordinates(val lon: Long, val lat: Long)
+data class Coordinates(val lon: Double, val lat: Double)
